@@ -1,6 +1,6 @@
 {
  "name": "uxcore-form",
- "version": "5.3.4",
+ "version": "5.4.0",
  "description": "form component for uxcore",
  "repository": "https://github.com/uxcore/uxcore-form.git",
  "author": "eternalsky",
@@ -18,7 +18,8 @@
   "chrome": "uxcore-tools run chrome",
   "browsers": "uxcore-tools run browsers",
   "saucelabs": "uxcore-tools run saucelabs",
-  "update": "uxcore-tools run update"
+  "update": "uxcore-tools run update",
+  "tnpm-update": "uxcore-tools run tnpm-update"
  },
  "bugs": {
   "url": "http://github.com/uxcore/uxcore-form/issues"
@@ -31,13 +32,15 @@
  ],
  "devDependencies": {
   "console-polyfill": "^0.2.2",
-  "enzyme": "^2.7.0",
+  "enzyme": "^3.0.0",
+  "enzyme-adapter-react-15": "^1.0.0",
   "es5-shim": "^4.5.8",
   "expect.js": "~0.3.1",
   "kuma-base": "1.x",
-  "react": "0.14.x",
-  "react-addons-test-utils": "0.14.x",
-  "react-dom": "0.14.x",
+  "react": "15.x",
+  "react-addons-test-utils": "15.x",
+  "react-dom": "15.x",
+  "react-test-renderer": "15.x",
   "uxcore-button": "^0.4.5",
   "uxcore-kuma": "*",
   "uxcore-tools": "0.2.x",
@@ -49,6 +52,7 @@
   "lie": "^3.1.0",
   "lodash": "^4.16.6",
   "object-assign": "^4.0.0",
+  "prop-types": "15.x",
   "uxcore-checkbox-group": "^1.2.0",
   "uxcore-const": "~0.0.1",
   "uxcore-date-form-field": "~0.8.0",
