@@ -1,9 +1,12 @@
 import expect from 'expect.js';
 import React from 'react';
-import { mount } from 'enzyme';
+import Enzyme, { mount } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-15';
 import FormField from 'uxcore-form-field';
 import FormRow from 'uxcore-form-row';
 import Form from '../src/Form';
+
+Enzyme.configure({ adapter: new Adapter() });
 
 /* eslint-disable react/no-string-refs */
 describe('FormRow', () => {
@@ -22,12 +25,12 @@ describe('FormRow', () => {
         <FormRow className="kuma-row-test" />
       </Form>
     );
-    expect(wrapper.find('.kuma-row-test')).to.have.length(1);
+    expect(wrapper.find('div.kuma-row-test')).to.have.length(1);
   });
   it('should be able to wrap FormField', () => {
     wrapper = mount(
       <Form>
-        <FormRow className="kuma-row-test">
+        <FormRow>
           <FormField jsxname="test" />
         </FormRow>
       </Form>
